requirements.txt
# Core
python-telegram-bot==21.9
fastapi==0.115.6
mangum==0.19.0
httpx==0.27.2

# Vector DB & Embeddings
pinecone==5.4.2
openai==1.58.1         # Used for both OpenRouter (LLM) and OpenAI (embeddings)

# Data parsing
python-frontmatter==1.1.0
pydantic==2.10.3
pydantic-settings==2.7.0

# Utilities
python-dotenv==1.0.1
tqdm==4.67.1

# Dev / testing
pytest==8.3.4
pytest-asyncio==0.25.0
ruff==0.8.6
